+++ src/main/kotlin/com/cycode/plugin/cli/models/scanResult/iac/IacScanResult.kt
@@ -1,9 +1,9 @@
-package com.cycode.plugin.cli.models.scanResult.sca
+package com.cycode.plugin.cli.models.scanResult.iac
 
 import com.cycode.plugin.cli.models.CliError
 import com.cycode.plugin.cli.models.scanResult.ScanResultBase
 
-data class ScaScanResult(
-    override val detections: List<ScaDetection>,
+data class IacScanResult(
+    override val detections: List<IacDetection>,
     override val errors: List<CliError>,
 ) : ScanResultBase